Added auto sync feature (fixes #45) (#140)
* added auto sync feature

* clean up , fix code climate

* clean up , fix code climate

* code climate conflict fixed
