+++ app/src/main/java/org/ole/planet/takeout/Data/realm_UserModel.java
@@ -11,7 +11,9 @@ import org.ole.planet.takeout.utilities.Utilities;
 import java.util.Map;
 import java.util.Set;
 
+import io.realm.Realm;
 import io.realm.RealmObject;
+import io.realm.RealmResults;
 import io.realm.annotations.PrimaryKey;
 
 public class realm_UserModel extends RealmObject {
@@ -192,4 +194,41 @@ public class realm_UserModel extends RealmObject {
             }
         }
     }
+
+    public static void populateUsersTable(JsonObject jsonDoc, Realm mRealm, SharedPreferences settings) {
+        try {
+            RealmResults<realm_UserModel> db_users = mRealm.where(realm_UserModel.class)
+                    .equalTo("id", jsonDoc.get("_id").getAsString())
+                    .findAll();
+            if (db_users.isEmpty()) {
+                realm_UserModel user = mRealm.createObject(realm_UserModel.class, jsonDoc.get("_id").getAsString());
+                insertIntoUsers(jsonDoc, user, settings);
+            }
+        } catch (Exception err) {
+            err.printStackTrace();
+        }
+    }
+
+
+    private static void insertIntoUsers(JsonObject jsonDoc, realm_UserModel user, SharedPreferences settings) {
+        user.set_rev(jsonDoc.get("_rev").getAsString());
+        user.setName(jsonDoc.get("name").getAsString());
+        user.setRoles("");
+        user.setUserAdmin(jsonDoc.get("isUserAdmin").getAsBoolean());
+        user.setJoinDate(jsonDoc.get("joinDate").getAsInt());
+        user.setFirstName(jsonDoc.get("firstName").getAsString());
+        user.setLastName(jsonDoc.get("lastName").getAsString());
+        user.setMiddleName(jsonDoc.get("middleName").getAsString());
+        user.setEmail(jsonDoc.get("email").getAsString());
+        user.setPhoneNumber(jsonDoc.get("phoneNumber").getAsString());
+        user.setPassword_scheme(jsonDoc.get("password_scheme").getAsString());
+        user.setIterations(jsonDoc.get("iterations").getAsString());
+        user.setDerived_key(jsonDoc.get("derived_key").getAsString());
+        user.setSalt(jsonDoc.get("salt").getAsString());
+        user.setDob(jsonDoc.get("birthDate") == null ? "" : jsonDoc.get("birthDate").getAsString());
+        user.setCommunityName(jsonDoc.get("communityName") == null ? "" : jsonDoc.get("communityName").getAsString());
+        user.addImageUrl(jsonDoc, settings);
+    }
+
+
 }
