+++ app/src/main/java/org/ole/planet/takeout/Data/realm_stepExam.java
@@ -1,4 +1,7 @@
 package org.ole.planet.takeout.Data;
+import com.google.gson.JsonObject;
+
+import io.realm.Realm;
 import io.realm.annotations.PrimaryKey;
 import io.realm.RealmObject;
 
@@ -58,4 +61,18 @@ public class realm_stepExam extends RealmObject {
     public void setName(String name) {
         this.name = name;
     }
+
+    public static void insertCourseStepsExams(String myCoursesID, String step_id, JsonObject exam, Realm mRealm) {
+        realm_stepExam myExam = mRealm.createObject(realm_stepExam.class, exam.get("_id").getAsString());
+        myExam.setStepId(step_id);
+        myExam.setCourseId(myCoursesID);
+        myExam.setName(exam.get("name").getAsString());
+        if (exam.has("passingPercentage"))
+            myExam.setPassingPercentage(exam.get("passingPercentage").getAsString());
+        if (exam.has("totalMarks"))
+            myExam.setPassingPercentage(exam.get("totalMarks").getAsString());
+        if (exam.has("questions"))
+          realm_examQuestion.insertExamQuestions(exam.get("questions").getAsJsonArray(), exam.get("_id").getAsString(), mRealm);
+    }
+
 }
