+++ app/src/main/java/org/ole/planet/takeout/SyncActivity.java
@@ -5,7 +5,6 @@ import android.content.DialogInterface;
 import android.content.SharedPreferences;
 import android.support.annotation.Nullable;
 import android.support.v7.app.AlertDialog;
-import android.util.Log;
 import android.view.View;
 import android.widget.ArrayAdapter;
 import android.widget.CompoundButton;
@@ -17,20 +16,23 @@ import com.afollestad.materialdialogs.MaterialDialog;
 
 import org.lightcouch.CouchDbProperties;
 import org.ole.planet.takeout.Data.realm_UserModel;
+import org.ole.planet.takeout.callback.SyncListener;
+import org.ole.planet.takeout.datamanager.DatabaseService;
+import org.ole.planet.takeout.service.SyncManager;
+import org.ole.planet.takeout.utilities.NotificationUtil;
 
 import java.net.URI;
 import java.util.ArrayList;
 import java.util.List;
 
 import io.realm.Realm;
-import io.realm.RealmConfiguration;
 import io.realm.RealmResults;
 
-public abstract class SyncActivity extends ProcessUserData {
-    private TextView syncDate;
-    private TextView intervalLabel;
-    private Spinner spinner;
-    private Switch syncSwitch;
+public abstract class SyncActivity extends ProcessUserData implements SyncListener {
+    public TextView syncDate;
+    public TextView intervalLabel;
+    public Spinner spinner;
+    public Switch syncSwitch;
     int convertedDate;
     public static final String PREFS_NAME = "OLE_PLANET";
     SharedPreferences settings;
@@ -42,23 +44,27 @@ public abstract class SyncActivity extends ProcessUserData {
 
     public void sync(MaterialDialog dialog) {
         // Check Autosync switch (Toggler)
+        spinner = (Spinner) dialog.findViewById(R.id.intervalDropper);
         syncSwitch = (Switch) dialog.findViewById(R.id.syncSwitch);
         intervalLabel = (TextView) dialog.findViewById(R.id.intervalLabel);
         syncSwitch.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
             @Override
             public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-                if (isChecked) {
-                    Log.e("MD: ", "Autosync is On");
-                    intervalLabel.setVisibility(View.VISIBLE);
-                    spinner.setVisibility(View.VISIBLE);
-                } else {
-                    Log.e("MD: ", "Autosync is Off");
-                    spinner.setVisibility(View.GONE);
-                    intervalLabel.setVisibility(View.GONE);
-                }
+                setSpinnerVisibility(isChecked);
             }
         });
+        syncSwitch.setChecked(settings.getBoolean("autoSync", false));
         dateCheck(dialog);
+    }
+
+    private void setSpinnerVisibility(boolean isChecked) {
+        if (isChecked) {
+            intervalLabel.setVisibility(View.VISIBLE);
+            spinner.setVisibility(View.VISIBLE);
+        } else {
+            spinner.setVisibility(View.GONE);
+            intervalLabel.setVisibility(View.GONE);
+        }
     }
 
     private void dateCheck(MaterialDialog dialog) {
@@ -73,7 +79,6 @@ public abstract class SyncActivity extends ProcessUserData {
         }
 
         // Init spinner dropdown items
-        spinner = (Spinner) dialog.findViewById(R.id.intervalDropper);
         syncDropdownAdd();
     }
 
@@ -95,24 +100,6 @@ public abstract class SyncActivity extends ProcessUserData {
         spinner.setAdapter(spinnerArrayAdapter);
     }
 
-
-    public void syncDatabase() {
-        Thread td = new Thread(new Runnable() {
-            public void run() {
-                try {
-                    realmConfig("_users");
-                    userTransactionSync(settings, mRealm, properties, progress_dialog);
-                    myLibraryTransactionSync();
-                } finally {
-                    if (mRealm != null) {
-                        mRealm.close();
-                    }
-                }
-            }
-        });
-        td.start();
-    }
-
     public void alertDialogOkay(String Message) {
         AlertDialog.Builder builder1 = new AlertDialog.Builder(this);
         builder1.setMessage(Message);
@@ -131,7 +118,8 @@ public abstract class SyncActivity extends ProcessUserData {
         this.settings = settings;
         this.context = context;
         AndroidDecrypter decrypt = new AndroidDecrypter();
-        realmConfig("_users");
+        mRealm = new DatabaseService(context).getRealmInstance();
+        ;
         if (mRealm.isEmpty()) {
             alertDialogOkay("Server not configured properly. Connect this device with Planet server");
             mRealm.close();
@@ -164,30 +152,6 @@ public abstract class SyncActivity extends ProcessUserData {
         return false;
     }
 
-
-    public void realmConfig(String dbName) {
-        Realm.init(context);
-        RealmConfiguration config = new RealmConfiguration.Builder()
-                .name(Realm.DEFAULT_REALM_NAME)
-                .deleteRealmIfMigrationNeeded()
-                .schemaVersion(4)
-                .build();
-        Realm.setDefaultConfiguration(config);
-        mRealm = Realm.getInstance(config);
-        properties = new CouchDbProperties()
-                .setDbName(dbName)
-                .setCreateDbIfNotExist(false)
-                .setProtocol(settings.getString("url_Scheme", "http"))
-                .setHost(settings.getString("url_Host", "192.168.2.1"))
-                .setPort(settings.getInt("url_Port", 3000))
-                .setUsername(settings.getString("url_user", ""))
-                .setPassword(settings.getString("url_pwd", ""))
-                .setMaxConnections(100)
-                .setConnectionTimeout(0);
-
-    }
-
-
     public void setUrlParts(String url, String password, Context context) {
         this.context = context;
         URI uri = URI.create(url);
@@ -208,13 +172,21 @@ public abstract class SyncActivity extends ProcessUserData {
         editor.putString("url_user", url_user);
         editor.putString("url_pwd", url_pwd);
         editor.commit();
+        SyncManager.getInstance().start(this);
+    }
+
+    @Override
+    public void onSyncStarted() {
         progress_dialog = new MaterialDialog.Builder(this)
                 .title("Syncing")
                 .content("Please wait")
                 .progress(true, 0)
                 .show();
-        syncDatabase();
     }
 
-
+    @Override
+    public void onSyncComplete() {
+        progress_dialog.dismiss();
+        NotificationUtil.cancellAll(this);
+    }
 }
