+++ app/src/main/java/org/ole/planet/takeout/callback/SyncListener.java
@@ -1,0 +1,7 @@
+package org.ole.planet.takeout.callback;
+
+public interface SyncListener {
+    void onSyncStarted();
+
+    void onSyncComplete();
+}
