+++ app/src/main/java/org/ole/planet/takeout/utilities/NotificationUtil.java
@@ -1,0 +1,52 @@
+package org.ole.planet.takeout.utilities;
+
+import android.app.Notification;
+import android.app.NotificationChannel;
+import android.app.NotificationManager;
+import android.app.PendingIntent;
+import android.content.Context;
+import android.content.Intent;
+import android.support.v4.app.NotificationCompat;
+import android.support.v4.app.NotificationManagerCompat;
+
+
+public class NotificationUtil {
+
+
+
+    public static void create(Context context, int smallIcon, String contentTitle, String contentText) {
+        NotificationManager manager =
+                (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
+
+        NotificationCompat.Builder a = new NotificationCompat.Builder(context, "11");
+
+       setChannel(manager);
+
+        Notification noti = a
+                .setContentTitle(contentTitle)
+                .setContentText(contentText)
+                .setSmallIcon(smallIcon)
+                .setProgress(0, 0, true)
+                .setAutoCancel(true).build();
+        manager.notify(111, noti);
+
+    }
+
+    public static void cancel(Context context, int id) {
+        NotificationManagerCompat nm = NotificationManagerCompat.from(context);
+        nm.cancel(id);
+    }
+
+    public static void cancellAll(Context context) {
+        NotificationManagerCompat nm = NotificationManagerCompat.from(context);
+        nm.cancelAll();
+    }
+
+    public static void setChannel(NotificationManager notificationManager) {
+        if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.O) {
+            int importance = NotificationManager.IMPORTANCE_LOW;
+            NotificationChannel notificationChannel = new NotificationChannel("11", "ole", importance);
+            notificationManager.createNotificationChannel(notificationChannel);
+        }
+    }
+}
