+++ app/src/main/res/layout/dialog_server_url_.xml
@@ -77,6 +77,7 @@
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:layout_alignParentLeft="true"
+                    android:layout_centerVertical="true"
                     android:text="Interval"
                     android:textColor="@color/dialog_sync_labels"
                     android:visibility="gone" />
